Split up authenticators into separate files
