+++ Sources/App/Framework/JWT/JWTError+TokenExpried.swift
@@ -1,0 +1,6 @@
+import JWT
+
+extension JWTError {
+    /// JWT Error thrown when the submitted token is expired.
+    static let tokenExpired = Self.claimVerificationFailure(name: "exp", reason: "expired")
+}
